refactor(direct-control): extract ready unit source
Move the ready-unit embedded App UI source into src/play/ready/unit.ts while keeping the public wrapper and command builder in the facade. Keep the move as source ownership only; this does not claim live-game runtime proof.

Record the ready atom status in the support modularization OpenSpec workstream, including the remaining ready-city source owner and the direct DRA correction-channel fallback rule.

Validation: bun run --cwd packages/civ7-direct-control test -- ready-unit-view.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
